Lab8SMA

+++ app/src/main/java/com/example/smartwallet/model/MonthlyExpenses.kt
@@ -1,5 +1,5 @@
 package com.example.smartwallet.model
 
-data class MonthlyExpenses(var month:String, val expenses: Int, val income: Int){
-    constructor(): this("",0,0)
+data class MonthlyExpenses(var month:String, val expenses: Double, val income: Double){
+    constructor(): this("",0.00,0.00)
 }
